feat: contact classes are no longer child classes
