+++ src/components/folder/DeleteFolder.tsx
@@ -56,7 +56,7 @@ const DeleteFolder = ({ folder }: DeleteFolderProps): JSX.Element => {
 				const filesToDelete = user.drive.files.filter((file) => file.path.startsWith(fullPath));
 				if (filesToDelete.length > 0) {
 					const filesResponse = await user.drive.delete(filesToDelete.map((file) => file.hash));
-					await user.fullContact.contact.deleteFiles(
+					await user.fullContact.files.deleteFiles(
 						filesToDelete.map((file) => file.id),
 						[],
 					);
