+++ src/lib/contact/contactClasses/folderContact.ts
@@ -1,0 +1,7 @@
+import Contact from '../contact'
+
+class ContactFolder {
+
+}
+
+export default ContactFolder;
