+++ src/lib/contact/contactClasses/manageContact.ts
@@ -1,0 +1,7 @@
+import Contact from '../contact'
+
+class ManageContact {
+
+}
+
+export default ManageContact;
